refactor(policy): split the custom-check policy types at the large-file bar
The depBump.schedule addition pushed policy.ts over the 500-line floor
(the self-guardrail flagged it). CustomCheckConfig + LintPolicy - the
custom-check seam's declarations - move to policy-checks.ts, re-exported
from policy.ts so every consumer import keeps working.
